// 11-directiveSignals/src/app/signals/interfaces/user.response..interface.ts
// Generated by https://quicktype.io

export interface SingleUserResponse {
  data:    User;
  support: Support;
}

export interface User {
  id:         number;
  email:      string;
  first_name: string;
  last_name:  string;
  avatar:     string;
}

export interface Support {
  url:  string;
  text: string;
}